example of protecting the delete comment - authorisation, authentication and sql injection ^^

+++ admin/includes/view_all_users.php
@@ -43,7 +43,7 @@
             echo "<td>{$user_email}</td>";
             echo "<td>{$user_role}</td>";
 
-            
+
 
             // $query = "SELECT * FROM categories WHERE cat_id = $post_category_id ";
             // $select_categories_id = mysqli_query($connection, $query);
@@ -83,14 +83,23 @@
 <?php
 // delete comment
 if (isset($_GET['delete'])) {
-    $the_user_id = $_GET['delete'];
 
-    $query = "DELETE FROM users WHERE user_id = $the_user_id ";
-    $delete_user_query = mysqli_query($connection, $query);
+    if (isset($_SESSION['user_role'])) {
 
 
-    // reloading
-    header("Location: users.php");
+        if ($_SESSION['user_role'] == 'admin') {
+
+
+            $the_user_id = mysqli_real_escape_string($connection, $_GET['delete']);
+
+            $query = "DELETE FROM users WHERE user_id = $the_user_id ";
+            $delete_user_query = mysqli_query($connection, $query);
+
+
+            // reloading
+            header("Location: users.php");
+        }
+    }
 }
